workflow-diagram: move identifyPlaceholder out of component
