+++ assets/js/workflow-diagram/util/placeholder.ts
@@ -1,4 +1,5 @@
 import { Flow } from '../types';
+import { WorkflowProps } from '../workflow-editor/store';
 
 // adds a placeholder node as child of the target
 // A node can only have one placeholder at a time
@@ -27,3 +28,39 @@ export const add = (_model: Flow.Model, parentNode: Flow.Node) => {
 };
 
 export const isPlaceholder = (node: Node) => node.placeholder;
+
+type Workflow = Pick<WorkflowProps, 'jobs' | 'edges' | 'triggers'>;
+
+// Identify placeholder nodes and return a new workflow model
+export const identify = (store: Workflow) => {
+  const { jobs, triggers, edges } = store;
+
+  const newJobs = jobs.map(item => {
+    if (!item.name && !item.body) {
+      return {
+        ...item,
+        placeholder: true,
+      };
+    }
+    return item;
+  });
+
+  const newEdges = edges.map(edge => {
+    const target = newJobs.find(({ id }) => edge.target_job_id === id);
+    if (target?.placeholder) {
+      return {
+        ...edge,
+        placeholder: true,
+      };
+    }
+    return edge;
+  });
+
+  const result = {
+    triggers,
+    jobs: newJobs,
+    edges: newEdges,
+  };
+
+  return result;
+};
